LayoutTests/http/tests/security/window-name-after-same-origin-main-frame-navigation-expected.txt

CONSOLE MESSAGE: line 9: window.name should equal "test" after a same-origin main frame navigation.
CONSOLE MESSAGE: line 4: window.name equals "test".
